Charge and scoring complete.

+++ shared/src/main/scala/models/board/mutation/UpdateSegment.scala
@@ -4,6 +4,6 @@ case class UpdateSegment(
   category: String,
   mutations: Seq[Mutation],
   combo: Option[Int] = None,
-  charge: Option[Int] = None,
+  charge: Option[Double] = None,
   scoreDelta: Option[Int] = None
 )
